Refactor code to handle diverse genus data types
The modified code now takes an argument for the type of genus data to analyze, such as abundance or distance diversity data, and computes measures like the Shannon index, Simpson index, and Pielou index accordingly. This change allows a broader range of data inputs and measures to be used, increasing the flexibility of the function. Accompanying this, the code for plotting the results has also been adjusted to handle different types of data and measures.